refact : 각 앨범 목록 페이지에서 사용하는 템플릿 분리 후 공통으로 Import
